Under construction

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,13 +1,17 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entites;
+using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
-using Restaurants.Domain.Repositories;
+using Restaurants.Domain.IRepository;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
-public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantRepository restaurantRepository ) : 
-    IRequestHandler<DeleteRestaurantCommand>
+public class DeleteRestaurantCommandHandler(
+    ILogger<DeleteRestaurantCommandHandler> logger, 
+    IRestaurantRepository restaurantRepository,
+    IAuthorisationService authorisationService) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -15,6 +19,9 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
         var restaurant = await restaurantRepository.GetRestaurantByIdAsync(request.Id);
         if (restaurant is null)
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+        
+        if (!authorisationService.Authorise(restaurant, ResourceOperation.Delete))
+            throw new ForbiddenException("User does not have permission to delete this restaurant");
             
         await restaurantRepository.DeleteRestaurantAsync(restaurant);
     }
